Moved CenteredImage
- makes more sense to have this component in its own folder, specifically for components made to help MDX authoring
