+++ components/content-helpers/CenteredImage.js
